// src/modules/pricing/enums/pricing-rule-type.enum.ts
export enum PricingRuleType {
  MARGIN_FLOOR = 'margin_floor',
  MAX_DISCOUNT = 'max_discount',
  QTY_BREAK = 'qty_break',
  LEAD_TIME = 'lead_time',
  CATALOG_UNIT = 'catalog_unit',
  INCLUDED_ALLOWANCE = 'included_allowance',
  CONDITIONAL_SURCHARGE = 'conditional_surcharge',
  FIXED_ADDER = 'fixed_adder',
  LABOR_HOURS = 'labor_hours',
  MARGIN_MARKUP = 'margin_markup',
  TAX = 'tax',
}